+++ python/src/worker/helpers/color_pool.py
