+++ python/src/worker/helpers/count_matrix.py
@@ -3,7 +3,7 @@ import datetime
 import os
 from datetime import timezone
 from logging import error, info
-from config import config
+from ..config import config
 import aws_xray_sdk as xray
 from aws_xray_sdk.core import xray_recorder
