+++ python/src/worker/helpers/find_cell_ids_in_same_hierarchy.py
