+++ python/src/worker/tasks/cluster_cells.py
@@ -2,10 +2,10 @@ import pandas as pd
 import json
 import requests
 import backoff
-from result import Result
-from helpers.color_pool import COLOR_POOL
-from config import config
-from tasks import Task
+from ..result import Result
+from ..helpers.color_pool import COLOR_POOL
+from ..config import config
+from ..tasks import Task
 from aws_xray_sdk.core import xray_recorder
 from natsort import natsorted
