+++ python/src/worker/tasks/differential_expression.py
@@ -1,15 +1,15 @@
 import json
-from config import config
-from result import Result
+from ..config import config
+from ..result import Result
 import pandas
 import requests
 import backoff
 from aws_xray_sdk.core import xray_recorder
 
-from helpers.s3 import get_cell_sets
-from helpers.find_cells_by_set_id import find_cells_by_set_id
-from helpers.find_cell_ids_in_same_hierarchy import find_cell_ids_in_same_hierarchy, find_all_cell_ids_in_cell_sets
-from tasks import Task
+from ..helpers.s3 import get_cell_sets
+from ..helpers.find_cells_by_set_id import find_cells_by_set_id
+from ..helpers.find_cell_ids_in_same_hierarchy import find_cell_ids_in_same_hierarchy, find_all_cell_ids_in_cell_sets
+from ..tasks import Task
 
 
 class DifferentialExpression(Task):
