+++ python/src/worker/tasks/list_genes.py
@@ -1,9 +1,9 @@
 import backoff
 import pandas as pd
-from result import Result
+from ..result import Result
 import requests
-from config import config
-from tasks import Task
+from ..config import config
+from ..tasks import Task
 import json
 from aws_xray_sdk.core import xray_recorder
